Crate canopy_derive

Attribute Macros

command

Mark a method as a command. This macro should be used to decorate methods in an impl block that uses the derive_commands macro. A number of optional arguments can be passed:

derive_commands

Derive an implementation of the CommandNode trait. This macro should be added to the impl block of a struct. All methods that are annotated with command are added as commands, with their doc comments as the command documentation.

Derive Macros

StatefulNode

Derive an implementation of the StatefulNode trait for a struct. The struct should have a self.state attribute of type NodeState.